Linkify headings inside of private articles
