+++ assets/js/analytics-events.tsx
@@ -2,7 +2,7 @@ import { h, render, Fragment } from "preact";
 import { createPortal } from "preact/compat";
 import { loggedInUser, PrivateLoginLink } from "./private";
 import { Alert } from "./components/alert";
-import { urlify } from "./urlify";
+import { AnchorLink, urlify } from "./components/anchor-link";
 import { Sidenav } from "./components/sidenav";
 
 interface AnalyticsEventAttribute {
@@ -16,34 +16,6 @@ interface AnalyticsEvent {
   previous_event_names?: string[];
   description: string;
   attributes: AnalyticsEventAttribute[];
-}
-
-function AnchorLink({
-  slug,
-  icon = String.fromCharCode(59851),
-}: {
-  slug: string;
-  icon?: string;
-}) {
-  const setRef = (node: HTMLElement | null) => {
-    if (node && document.location.hash.slice(1) === slug) {
-      setTimeout(() => node.scrollIntoView(), 0);
-    }
-  };
-
-  return (
-    // See: https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/issues/423
-    // eslint-disable-next-line jsx-a11y/anchor-has-content
-    <a
-      ref={setRef}
-      className="anchorjs-link"
-      aria-label="Anchor"
-      data-anchorjs-icon={icon}
-      id={slug}
-      href={`#${slug}`}
-      style={{ font: "1em / 1 anchorjs-icons", paddingLeft: "0.375em" }}
-    />
-  );
 }
 
 function Example({
