--- assets/js/urlify.tsx
@@ -1,5 +1,0 @@
-import Anchor from "anchor-js";
-
-const anchor = new Anchor();
-
-export const urlify = anchor.urlify.bind(anchor);
